Move parts of root index.tsx to App.tsx
